+++ scm-plugins/scm-hg-plugin/src/test/java/sonia/scm/repository/spi/HgBranchCommandTest.java
@@ -25,7 +25,6 @@
 package sonia.scm.repository.spi;
 
 import com.aragost.javahg.commands.PullCommand;
-import com.google.inject.util.Providers;
 import org.junit.Before;
 import org.junit.Test;
 import sonia.scm.repository.Branch;
@@ -46,7 +45,7 @@ public class HgBranchCommandTest extends AbstractHgCommandTestBase {
   @Before
   public void initWorkingCopyFactory() {
 
-    workingCopyFactory = new SimpleHgWorkingCopyFactory(new NoneCachingWorkingCopyPool(new WorkdirProvider())) {
+    workingCopyFactory = new SimpleHgWorkingCopyFactory(new NoneCachingWorkingCopyPool(new WorkdirProvider(repositoryLocationResolver))) {
       @Override
       public void configure(PullCommand pullCommand) {
         // we do not want to configure http hooks in this unit test